docs: moved provider options docs snippets
