+++ apps/docs-snippets/src/guide/provider/provider.test.ts
@@ -1,13 +1,5 @@
-import { Provider, ScriptTransactionRequest, sleep, WalletUnlocked, Address } from 'fuels';
+import { Provider, ScriptTransactionRequest, WalletUnlocked, Address } from 'fuels';
 import { launchTestNode } from 'fuels/test-utils';
-
-async function fetchSomeExternalCredentials() {
-  return Promise.resolve('credential');
-}
-
-function decorateResponseWithCustomLogic(response: Response) {
-  return response;
-}
 
 /**
  * @group node
@@ -44,97 +36,6 @@ describe('Provider', () => {
     expect(consensusParameters).toBeDefined();
     expect(consensusParameters).toBeInstanceOf(Object);
     expect(balances).toEqual([]);
-  });
-
-  test('options: requestMiddleware', async () => {
-    using launched = await launchTestNode();
-
-    const FUEL_NETWORK_URL = launched.provider.url;
-
-    // #region options-requestMiddleware
-    // synchronous request middleware
-    await Provider.create(FUEL_NETWORK_URL, {
-      requestMiddleware: (request: RequestInit) => {
-        request.credentials = 'omit';
-
-        return request;
-      },
-    });
-
-    // asynchronous request middleware
-    await Provider.create(FUEL_NETWORK_URL, {
-      requestMiddleware: async (request: RequestInit) => {
-        const credentials = await fetchSomeExternalCredentials();
-        request.headers ??= {};
-        (request.headers as Record<string, string>).auth = credentials;
-
-        return request;
-      },
-    });
-    // #endregion options-requestMiddleware
-  });
-
-  it('options: timeout', async () => {
-    using launched = await launchTestNode();
-
-    const FUEL_NETWORK_URL = launched.provider.url;
-
-    // #region options-timeout
-    await Provider.create(FUEL_NETWORK_URL, {
-      timeout: 30000, // will abort if request takes 30 seconds to complete
-    });
-    // #endregion options-timeout
-  });
-
-  it('options: retryOptions', async () => {
-    using launched = await launchTestNode();
-
-    const FUEL_NETWORK_URL = launched.provider.url;
-
-    // #region options-retryOptions
-    await Provider.create(FUEL_NETWORK_URL, {
-      retryOptions: {
-        maxRetries: 5,
-        baseDelay: 100,
-        backoff: 'linear',
-      },
-    });
-    // #endregion options-retryOptions
-  });
-
-  it('options: fetch', async () => {
-    using launched = await launchTestNode();
-
-    const FUEL_NETWORK_URL = launched.provider.url;
-
-    // #region options-fetch
-    await Provider.create(FUEL_NETWORK_URL, {
-      fetch: async (url: string, requestInit: RequestInit | undefined) => {
-        // do something
-        await sleep(100);
-
-        // native fetch
-        const response = await fetch(url, requestInit);
-
-        const updatedResponse = decorateResponseWithCustomLogic(response);
-
-        return updatedResponse;
-      },
-    });
-    // #endregion options-fetch
-  });
-
-  it('options: resourceCacheTTL', async () => {
-    using launched = await launchTestNode();
-
-    const FUEL_NETWORK_URL = launched.provider.url;
-    // #region options-cache-utxo
-    const provider = await Provider.create(FUEL_NETWORK_URL, {
-      resourceCacheTTL: 5000, // cache resources (Coin's and Message's) for 5 seconds
-    });
-    // #endregion options-cache-utxo
-
-    expect(provider).toBeDefined();
   });
 
   it('fetches the base asset ID', async () => {
